+++ src/main/scala/is/hail/expr/Relational.scala
@@ -4,7 +4,7 @@ import is.hail.HailContext
 import is.hail.annotations._
 import is.hail.asm4s.FunctionBuilder
 import is.hail.expr.ir._
-import is.hail.keytable.KTLocalValue
+import is.hail.table.KTLocalValue
 import is.hail.methods.Aggregators
 import is.hail.sparkextras._
 import is.hail.rvd._
