+++ src/main/scala/is/hail/io/CassandraConnector.scala
@@ -5,7 +5,7 @@ import java.nio.ByteBuffer
 import com.datastax.driver.core.querybuilder.QueryBuilder
 import com.datastax.driver.core.{Cluster, DataType, Session, TableMetadata, Row => CassRow}
 import is.hail.expr._
-import is.hail.keytable.Table
+import is.hail.table.Table
 import is.hail.utils._
 import org.apache.spark.sql.Row
 import org.json4s._
