+++ src/main/scala/is/hail/io/SolrConnector.scala
@@ -3,7 +3,7 @@ package is.hail.io
 import java.util
 
 import is.hail.expr._
-import is.hail.keytable.Table
+import is.hail.table.Table
 import is.hail.utils._
 import is.hail.variant.MatrixTable
 import org.apache.solr.client.solrj.impl.{CloudSolrClient, HttpSolrClient}
