+++ src/main/scala/is/hail/methods/Aggregators.scala
@@ -5,7 +5,7 @@ import java.io.{ObjectInputStream, ObjectOutputStream}
 import is.hail.HailContext
 import is.hail.annotations.{Annotation, RegionValue, RegionValueBuilder, UnsafeRow}
 import is.hail.expr._
-import is.hail.keytable.KTLocalValue
+import is.hail.table.KTLocalValue
 import is.hail.stats._
 import is.hail.utils._
 import is.hail.variant._
