+++ src/test/scala/is/hail/expr/TableIRSuite.scala
@@ -1,7 +1,7 @@
 package is.hail.expr
 
 import is.hail.SparkSuite
-import is.hail.keytable.Table
+import is.hail.table.Table
 import is.hail.expr.ir
 import org.apache.spark.sql.Row
 import org.testng.annotations.Test
