+++ src/test/scala/is/hail/methods/LinearRegressionSuite.scala
@@ -4,7 +4,7 @@ import is.hail.SparkSuite
 import is.hail.TestUtils._
 import is.hail.annotations.Annotation
 import is.hail.expr.{TFloat64, TString}
-import is.hail.keytable.Table
+import is.hail.table.Table
 import is.hail.utils._
 import is.hail.variant.Variant
 import org.testng.annotations.Test
