+++ src/test/scala/is/hail/methods/LogisticRegressionSuite.scala
@@ -3,7 +3,7 @@ package is.hail.methods
 import is.hail.SparkSuite
 import is.hail.annotations.{Annotation, Querier}
 import is.hail.expr.{TBoolean, TFloat64}
-import is.hail.keytable.Table
+import is.hail.table.Table
 import is.hail.utils._
 import is.hail.variant.Variant
 import org.testng.annotations.Test
