+++ src/test/scala/is/hail/variant/GenomeReferenceSuite.scala
@@ -3,7 +3,7 @@ package is.hail.variant
 import java.io.FileNotFoundException
 
 import is.hail.expr.{TInterval, TLocus, TStruct, TVariant}
-import is.hail.keytable.Table
+import is.hail.table.Table
 import is.hail.utils.Interval
 import is.hail.{SparkSuite, TestUtils}
 import org.apache.spark.sql.Row
