+++ src/test/scala/is/hail/vds/JoinSuite.scala
@@ -3,7 +3,7 @@ package is.hail.vds
 import is.hail.SparkSuite
 import is.hail.annotations.UnsafeRow
 import is.hail.expr.{TStruct, TVariant}
-import is.hail.keytable.Table
+import is.hail.table.Table
 import is.hail.variant.{GenomeReference, Variant, MatrixTable, VariantDataset}
 import is.hail.utils._
